Updated values to be i16 and reduced the size of Stat struct
Times:
        2.809 s | 5.1894 GB/s
        2.817 s | 5.1764 GB/s
        2.820 s | 5.1702 GB/s
        2.843 s | 5.1284 GB/s
        2.846 s | 5.1225 GB/s
Best  : 2.809 s | 5.1894 GB/s
Mean  : 2.827 s | 5.1572 GB/s
Median: 2.820 s | 5.1702 GB/s
Worst : 2.846 s | 5.1225 GB/s
BRC   : 2.826 s | 5.1582 GB/s
{"Length":15654537897,"Nanoseconds":[2809478000,2816517500,2819894100,2842876100,2846169900]}

+++ src/BRCMap.zig
@@ -106,7 +106,7 @@ pub fn BRCMapUnmanaged(comptime capacity: comptime_int) type {
             unreachable;
         }
 
-        pub fn addOrUpdate(self: *Self, key: []const u8, value: i32) void {
+        pub fn addOrUpdate(self: *Self, key: []const u8, value: i16) void {
             switch (self.findKeyIndex(key)) {
                 .found => |index| {
                     std.debug.assert(self.keys[index].isNotEmpty());
@@ -121,7 +121,7 @@ pub fn BRCMapUnmanaged(comptime capacity: comptime_int) type {
             }
         }
 
-        pub fn addOrUpdateCloned(self: *Self, gpa: std.mem.Allocator, key: []const u8, value: i32) !void {
+        pub fn addOrUpdateCloned(self: *Self, gpa: std.mem.Allocator, key: []const u8, value: i16) !void {
             switch (self.findKeyIndex(key)) {
                 .found => |index| {
                     std.debug.assert(self.keys[index].isNotEmpty());
