+++ src/root.zig
@@ -66,22 +66,22 @@ test LineSplitter {
         both_null = (std_item == null) and (new_item == null);
     }
 }
-pub fn brcIntParse(str: []const u8) i32 {
+pub fn brcIntParse(str: []const u8) i16 {
     const isNegative: bool = str[0] == '-';
-    const isNegativeInt: i32 = @intFromBool(isNegative);
-    const isPositiveInt: i32 = @intFromBool(!isNegative);
+    const isNegativeInt: i16 = @intFromBool(isNegative);
+    const isPositiveInt: i16 = @intFromBool(!isNegative);
     const arr: []const u8 = str[@intFromBool(isNegative)..];
     return ((-1 * isNegativeInt) + isPositiveInt) * // sign
-        (@as(i32, @intCast(arr[arr.len - 1] - '0')) + // 1s place
-            @as(i32, @intCast(arr[arr.len - 3] - '0')) * 10 + // 10s place
-            if (arr.len == 4) @as(i32, @intCast(arr[arr.len - 4] - '0')) * 100 else 0); // 100s place
+        (@as(i16, @intCast(arr[arr.len - 1] - '0')) + // 1s place
+            @as(i16, @intCast(arr[arr.len - 3] - '0')) * 10 + // 10s place
+            if (arr.len == 4) @as(i16, @intCast(arr[arr.len - 4] - '0')) * 100 else 0); // 100s place
 }
 test brcIntParse {
     const min: comptime_int = -999;
     const max: comptime_int = 999;
 
     var buf: [64]u8 = undefined;
-    var i: i32 = min;
+    var i: i16 = min;
     @memset(buf[0..], 0);
     while (i <= max) : (i += 1) {
         const f: f128 = @as(f128, @floatFromInt(i)) / 10.0;
